+++ .tests/cypress/_integration/mini-cart/remove-from-cart-simple.js
@@ -1,0 +1,59 @@
+describe('Simple product', function() {
+  it('Visits product', () => {
+    cy.visit('/joust-duffle-bag')
+    cy.get('.breadcrumbs__list').should('be.visible')
+  })
+
+  it('Check visiblility of content', () => {
+    cy.get('.product-view__main-details').should('be.visible')
+    cy.get('[data-ui-id=page-title-wrapper]').should('be.visible')
+    cy.get('.product-view__information').should('be.visible')
+    cy.get('.product-view__sku').should('be.visible')
+    cy.get('#product-price-1').should('be.visible')
+    cy.get('.quantity-update').should('be.visible')
+    cy.get('#product-addtocart-button').should('be.visible')
+    cy.get('.product-view__extra-actions').should('be.visible')
+    cy.get('[data-testid=product-tab__title]')
+      .contains('Details')
+      .should('be.visible')
+    cy.get('[data-testid=product-tab__title]')
+      .contains('Reviews')
+      .should('be.visible')
+  })
+
+  it('Add product to cart', () => {
+    cy.get('#product-addtocart-button')
+      .first()
+      .click()
+  })
+
+  it('Check if mini-cart is not empty', () => {
+    cy.get('[data-testid=minicart-link]').click()
+    cy.get('#minicart-content-wrapper').should('be.visible')
+    cy.contains('You have no items in your shopping cart.').should(
+      'not.be.visible'
+    )
+  })
+
+  it('Delate item form cart', () => {
+    cy.get('[data-testid=delete-item-link]').click()
+    // TODO: Why is this chained that way?
+    cy.get('.modal-inner-wrap .modal-footer').then(() => {
+      cy.get('.modal-footer button')
+        .eq(1)
+        .click()
+      cy.go('forward')
+      cy.server()
+      cy.request('/checkout/cart/').as('cartStatus')
+      cy.request('/customer/section/load/?sections=cart*').as('cartSection')
+      cy.route('/checkout/sidebar/removeItem/').as('removeItem')
+      cy.wait('@removeItem')
+    })
+  })
+
+  it('Is cart empty', () => {
+    cy.get('[data-testid=minicart-link]').click()
+    cy.get('#minicart-content-wrapper').should('be.visible')
+    cy.contains('You have no items in your shopping cart.').should('be.visible')
+  })
+})