+++ Magento_Theme/web/js/scroll-top.babel.js
@@ -1,0 +1,8 @@
+'use strict';
+define(() => {
+  return (options, element) => {
+    element.addEventListener('click', () => {
+      window.scrollTo(0, 0);
+    })
+  }
+})